bootcamp-2/requirements.txt
asyncpg
sqlmodel
pydantic-settings